Update README.md
Greetings everyone thats checking out my github profile. This is activity 3 of my colleague Mr. Criztian Tuplano's bootcamp, and this
is my output for activity 3 of the bootcamp. 
It is presently only the second week of the boot camp and we have started out with CSS and its concepts.
I don't know if I understood the concept of HTML heading class and id referenced in CSS colors. (I am not even sure if that's what you call these concepts)

So I am just going to try to and explain to the best of my ability, in the most concise way and manner that I can.
I also need to record a video after writing this  README.md, since the main instruction for this activity is to record a video explaining the concept of
HTML heading class and id referenced in CSS colors.

First of all, I learned that I the <br> and </br> tag is used in order to separate text from one another, 
forming a small margin between them, which I used in order to make my format my text so that it looks better, 
next I used align="center" in order to center my text and have it look better.

Now my interpretration of HTML heading class is that it is used in order to be able to declare a type of class for my HTML headings, in terms of my code it is
to define the value of name of the slasher horror villain competitor such as Michael Myers. And the HTML heading id is used in order to be able to change the 
color of the heading despite it not having a class and CSS thinking it's group within the same class since it uses the same heading such as H1.
I am not sure if this is right but my interpretation of HTML id when linked to a CSS file is that it takes priority over other parts of HTML tags like important 
and final.

As for the CSS well you can just declare the heading you want to be stylized using for example H1 with the color that you want it to be.  Though if you want to change that color even though it classified by the same HTML heading you can use the #syntax along with the class name of the HTML heading class for example #group-1 just which is the value attribute of my HTML heading class tag, this will have CSS prioritize this command rather than
the regular H1 color command.

Finally I decided to make a winner and loser color command in my CSS file, which I later overrode thanks to  the class attribute name HTML heading class syntax 
which is in this case was .group-1.winner  along with .group-2.winner.

Thank you so much for reading this long explanation about my code. I hoped I was able to explain it properly, in a very clear and concise manner.
In order to use and improve this repository for yourself (remember use and improve) and not plagiarize. Simply use git clone and clone this to your
local machine, then open up your text editor, which I am assuming in this case is VScode, and use live server to see the output of my HTML and CSS code.


And if you ever do, please credit me as this is my original work.